# Conflicts:
#	composer.json
#	composer.lock
#	coverage/BankModel.php.html
#	coverage/BaseModel.php.html
#	coverage/UserModel.php.html
#	coverage/dashboard.html
#	coverage/index.html
#	phpunit.xml
#	tests/BankModelTest.php
#	tests/UserModelTest.php

+++ models/BankModel.php
@@ -4,8 +4,11 @@ require_once 'BaseModel.php';
 
 class BankModel extends BaseModel {
 
-    public function findBankById($id) {
-        $sql = 'SELECT * FROM banks WHERE id = '.$id;
+    public function findBankById($id)
+    {
+        if(!is_numeric($id)) return 'error';
+        
+        $sql = 'SELECT * FROM banks WHERE id = ' . $id;
         $bank = $this->select($sql);
 
         return $bank;
@@ -17,10 +20,19 @@ class BankModel extends BaseModel {
      * @return mixed
      */
     public function deleteBankById($id) {
-        $subString = substr($id,36,-38);
-        $result = base64_decode($subString);
-        $sql = "DELETE FROM banks WHERE MD5(banks.id) = '" . md5($result) . "'";
-        return $this->delete($sql);
+        if(is_numeric($id)){    
+            if(is_float($id)){
+                return false;
+            }   
+            else{
+                $sql = 'DELETE FROM users WHERE id = '.$id;
+                return $this->delete($sql);
+            }   
+           
+        } 
+        else{
+            return false;
+        }
     }
     /**
      * Update bank
